+++ ProxyChannel.cxx
@@ -74,7 +74,7 @@ void PortRange::LoadConfig(const char *sec, const char *setting, const char *def
 {
 	PStringArray cfgs = GkConfig()->GetString(sec, setting, def).Tokenise(",.:-/'", FALSE);
 	if (cfgs.GetSize() >= 2) // no such a setting in config
-		port = minport = cfgs[0].AsUnsigned(), maxport = cfgs[1].AsUnsigned();
+		port = minport = (WORD)cfgs[0].AsUnsigned(), maxport = (WORD)cfgs[1].AsUnsigned();
 	else
 		port = 0;
 	PTRACE_IF(2, port, setting << ": " << minport << '-' << maxport);
